+++ packages/api/src/functions/remove.ts
@@ -1,7 +1,7 @@
 import type { Context, OptionalId } from '../types'
 import type { CollectionDocument, Filters } from './types'
-import { unsafe } from '@sonata-api/common'
-import { traverseDocument } from '../collection'
+import { left, unsafe } from '@sonata-api/common'
+import { traverseDocument, cascadingRemove } from '../collection'
 
 export const remove = <TDocument extends CollectionDocument<OptionalId<any>>>() => async <TContext>(payload: {
   filters: Filters<TDocument>
@@ -10,11 +10,22 @@ export const remove = <TDocument extends CollectionDocument<OptionalId<any>>>() 
   : never
 ) => {
   if( !payload.filters._id ) {
-    throw new Error('you must pass an _id as filter')
+    return left({
+      message: 'you must pass an _id as filter'
+    })
   }
 
-
-  return context.model.findOneAndDelete(unsafe(await traverseDocument(payload.filters, context.description, {
+  const filters = unsafe(await traverseDocument(payload.filters, context.description, {
     autoCast: true
-  })))
+  }))
+
+  const target = await context.model.findOne(filters)
+  if( !target ) {
+    return left({
+      message: 'target not found'
+    })
+  }
+
+  await cascadingRemove(target, context.description)
+  return context.model.findOneAndDelete(filters)
 }